refactor(web): extract pure free functions from App.tsx into src/lib
Phase 2 of issue #642 (plan comment 5160203979): move the backend/
ground-config free functions and types, the format helpers, and the
Richardson-extrapolation/color-blend math helpers out of App.tsx into
dedicated src/lib modules (backends.ts, ground.ts, format.ts, math.ts).
Verbatim code move — no logic changes, no reformatting; comments travel
with their definitions. Vec3/dot3/cross3 stay in App.tsx: Vec3 is used
by the component-side PROJECTIONS table, so it doesn't move cleanly on
its own.

The four Phase-1 test files now import directly from the lib modules
instead of re-exported names on App.tsx; only their import statements
changed, assertions are byte-identical.
